refactor: auto-split test_visualization_vr_backends.py
