+++ tests/unit/test_unreal_integration/visualization_vr_backends/test_create_viewer.py
@@ -1,0 +1,81 @@
+"""Unit tests for visualization, VR interaction, and viewer backends.
+
+TDD tests for the remaining Unreal Engine integration components.
+"""
+
+from __future__ import annotations
+
+import numpy as np
+import pytest
+from src.unreal_integration.data_models import (
+    ForceVector,
+    Quaternion,
+    SwingMetrics,
+    TrajectoryPoint,
+    Vector3,
+)
+from src.unreal_integration.mesh_loader import (
+    LoadedMesh,
+    MeshFace,
+    MeshVertex,
+)
+from src.unreal_integration.viewer_backends import (
+    BackendType,
+    CameraState,
+    MockBackend,
+    ViewerConfig,
+    create_viewer,
+)
+from src.unreal_integration.visualization import (
+    ForceVectorRenderer,
+    HUDDataProvider,
+    RenderData,
+    TrajectoryRenderer,
+    VisualizationConfig,
+    VisualizationType,
+)
+from src.unreal_integration.vr_interaction import (
+    VRControllerHand,
+    VRControllerState,
+    VRHeadsetState,
+    VRInteractionManager,
+    VRLocomotionMode,
+)
+
+# ============================================================================
+# Visualization Tests
+# ============================================================================
+
+
+# ============================================================================
+# VR Interaction Tests
+# ============================================================================
+
+
+# ============================================================================
+# Viewer Backend Tests
+# ============================================================================
+
+
+# Should not raise
+
+
+class TestCreateViewer:
+    """Tests for create_viewer factory function."""
+
+    def test_create_mock_viewer(self) -> None:
+        """Test creating mock viewer."""
+        viewer = create_viewer("mock")
+        assert isinstance(viewer, MockBackend)
+
+    def test_create_viewer_with_config(self) -> None:
+        """Test creating viewer with custom config."""
+        config = ViewerConfig(width=800, height=600)
+        viewer = create_viewer("mock", config=config)
+        assert viewer.config.width == 800
+        assert viewer.config.height == 600
+
+    def test_create_unsupported_viewer(self) -> None:
+        """Test creating unsupported viewer raises error."""
+        with pytest.raises(ValueError):
+            create_viewer("nonexistent_backend")
