+++ tests/unit/test_unreal_integration/visualization_vr_backends/test_render_data.py
@@ -1,0 +1,86 @@
+"""Unit tests for visualization, VR interaction, and viewer backends.
+
+TDD tests for the remaining Unreal Engine integration components.
+"""
+
+from __future__ import annotations
+
+import numpy as np
+import pytest
+from src.unreal_integration.data_models import (
+    ForceVector,
+    Quaternion,
+    SwingMetrics,
+    TrajectoryPoint,
+    Vector3,
+)
+from src.unreal_integration.mesh_loader import (
+    LoadedMesh,
+    MeshFace,
+    MeshVertex,
+)
+from src.unreal_integration.viewer_backends import (
+    BackendType,
+    CameraState,
+    MockBackend,
+    ViewerConfig,
+    create_viewer,
+)
+from src.unreal_integration.visualization import (
+    ForceVectorRenderer,
+    HUDDataProvider,
+    RenderData,
+    TrajectoryRenderer,
+    VisualizationConfig,
+    VisualizationType,
+)
+from src.unreal_integration.vr_interaction import (
+    VRControllerHand,
+    VRControllerState,
+    VRHeadsetState,
+    VRInteractionManager,
+    VRLocomotionMode,
+)
+
+# ============================================================================
+# Visualization Tests
+# ============================================================================
+
+
+# ============================================================================
+# VR Interaction Tests
+# ============================================================================
+
+
+# ============================================================================
+# Viewer Backend Tests
+# ============================================================================
+
+
+# Should not raise
+
+
+class TestRenderData:
+    """Tests for RenderData."""
+
+    def test_create_render_data(self) -> None:
+        """Test render data creation."""
+        data = RenderData(
+            visualization_type=VisualizationType.FORCE_ARROW,
+            vertices=np.array([[0, 0, 0], [1, 0, 0]]),
+        )
+        assert data.visualization_type == VisualizationType.FORCE_ARROW
+        assert len(data.vertices) == 2
+
+    def test_render_data_to_dict(self) -> None:
+        """Test render data serialization."""
+        data = RenderData(
+            visualization_type=VisualizationType.TRAJECTORY_LINE,
+            vertices=np.array([[0, 0, 0], [1, 1, 1]]),
+            colors=np.array([[1, 0, 0, 1], [0, 1, 0, 1]]),
+            metadata={"point_count": 2},
+        )
+        d = data.to_dict()
+        assert d["type"] == "trajectory_line"
+        assert len(d["vertices"]) == 2
+        assert d["metadata"]["point_count"] == 2
